Time: 0 ms (100%), Space: 28 MB (91.3%) - LeetHub

+++ 0169-majority-element/0169-majority-element.cpp
@@ -2,8 +2,8 @@ class Solution {
 public:
     int majorityElement(vector<int>& nums) {
         int element ;
-        long long  count = 0;
-        long long  count1 = 0;
+        int count = 0;
+        int count1 = 0;
         for(int i = 0 ; i < nums.size(); i++){
             if(count == 0){
                 count = 1;
